Add support for KDE Neon
The package ecosystem of KDE Neon is virtually identical to that of Ubuntu.
The release codenames and version numbers also reflect those of Ubuntu.
For the purposes of open-eid, there should be no significant differences
between the two distributions; hence, this patch adds a bit of simple
forwarding logic that detects KDE Neon, prints a warning about the lack
of official support, and hands the execution over to the Ubuntu branch.

Tested against a (mostly) fresh install of KDE Neon Focal 20.04, AMD64.

Signed-off-by: Pavel Kirienko <pavel.kirienko@gmail.com>